# exam-activities/movies-api-without-security/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peliculas api",
        "description": "Esta es la documentación de la API generada automáticamente con Swagger",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user and returns the JWT token for authentication",
                "operationId": "392091b67067bb84702403a281d29dff",
                "requestBody": {
                    "description": "User registration data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered and JWT token returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "JWT token for authentication",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user and get JWT token",
                "description": "Logs in the user and returns a JWT token for authentication",
                "operationId": "31dd365bddb46a996c2350b0a1fb7381",
                "requestBody": {
                    "description": "User login credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JWT token returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "JWT token for authentication",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access, invalid credentials"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/api/directors": {
            "get": {
                "tags": [
                    "Director"
                ],
                "summary": "Obtain the directors list",
                "description": "Returns the list of directors",
                "operationId": "128351fac0739f196796463e72a5a93a",
                "responses": {
                    "200": {
                        "description": "List of directors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Director"
                ],
                "summary": "Store a new director",
                "description": "Store a new director in the database",
                "operationId": "d0116341742a81504221cd1bd89aa0b5",
                "requestBody": {
                    "description": "Director data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellidos"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "First name of the director",
                                        "type": "string"
                                    },
                                    "apellidos": {
                                        "description": "Last name of the director",
                                        "type": "string"
                                    },
                                    "peliculas": {
                                        "description": "List of movie IDs related to the director",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Director created successfully"
                    },
                    "400": {
                        "description": "Invalid data provided"
                    }
                }
            }
        },
        "/api/directors/{id}": {
            "get": {
                "tags": [
                    "Director"
                ],
                "summary": "Get a specific director",
                "description": "Returns the details of a specific director by ID",
                "operationId": "4df4ad294227fd9b19cee83dec478ec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the director",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Director details"
                    },
                    "404": {
                        "description": "Director not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Director"
                ],
                "summary": "Update an existing director",
                "description": "Update the details of an existing director by ID",
                "operationId": "fd1aa0206c40645e8b1b0ae0265e7ec9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the director to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated director data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "description": "First name of the director",
                                        "type": "string"
                                    },
                                    "apellidos": {
                                        "description": "Last name of the director",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Director updated successfully"
                    },
                    "400": {
                        "description": "Invalid data provided"
                    },
                    "404": {
                        "description": "Director not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Director"
                ],
                "summary": "Delete a director",
                "description": "Delete a director by ID",
                "operationId": "f28c59230c7b92ee70dfa97500e957cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the director to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Director deleted successfully"
                    },
                    "404": {
                        "description": "Director not found"
                    }
                }
            }
        },
...
    },
...
}